// src/app/models/reports/acumuladoreal.ts
export interface IAcumuladoReal {
  descripcion:          string;
  idAutoLinea:          number;
  autoLinea:            string;
  enero:                number;
  eneroPerc:            number;
  febrero:              number;
  febreroPerc:          number;
  marzo:                number;
  marzoPerc:            number;
  abril:                number;
  abrilPerc:            number;
  mayo:                 number;
  mayoPerc:             number;
  junio:                number;
  junioPerc:            number;
  julio:                number;
  julioPerc:            number;
  agosto:               number;
  agostoPerc:           number;
  septiembre:           number;
  septiembrePerc:       number;
  octubre:              number;
  octubrePerc:          number;
  noviembre:            number;
  noviembrePerc:        number;
  diciembre:            number;
  diciembrePerc:        number;
  orderid:              number;
  totalAnual:           number;
  totalAnualPerc:       number;
  idEstadoResultadosI:  number;
  numeroCuenta:         string;
  titulo:               string;
  cantidadPrespuesto:   number;
  variacion:            number;
  cantidad:             number;
  percentVariacion:     number;
}